UPDATE: CLO-5365 - junit tests for SimpleSequence
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk-4-1@16877 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.sequence/src/org/jetel/sequence/SimpleSequence.java
@@ -92,7 +92,7 @@ public class SimpleSequence extends AbstractSequence implements Sequence {
      * @param numCachedValues	how many values should be cached (reduces IO but consumes some of the 
      * available values between object reusals)
      */
-    public SimpleSequence(String id, TransformationGraph graph, String sequenceName, String filename, long start, int step, int numCachedValues) {
+    public SimpleSequence(String id, TransformationGraph graph, String sequenceName, String filename, long start, long step, long numCachedValues) {
         super(id, graph, sequenceName);
         this.filename=filename;
         this.start=start;
